Extract CommandArgumentsBuilder
