+++ src/main/groovy/com/commercehub/gradle/cucumber/CucumberRunner.groovy
@@ -20,10 +20,6 @@ import java.util.jar.JarInputStream
  * Created by jgelais on 6/16/15.
  */
 class CucumberRunner {
-    private static final String PLUGIN = '--plugin'
-    private static final String TILDE = '~'
-    private static final String TAGS = '--tags '
-
     private static final String CUCUMBER_MAIN_NEW = 'io.cucumber.core.cli.Main'
     private static final String CUCUMBER_MAIN_OLD = 'cucumber.api.cli.Main'
 
@@ -71,15 +67,7 @@ class CucumberRunner {
                 File consoleErrLogFile = new File(resultsDir, "${featureName}-err.log")
                 File junitResultsFile = new File(resultsDir, "${featureName}.xml")
 
-                List<String> args = []
-                applyGlueArguments(args)
-                applyPluginArguments(args, resultsFile, junitResultsFile)
-                applyDryRunArguments(args)
-                applyMonochromeArguments(args)
-                applyStrictArguments(args)
-                applyTagsArguments(args)
-                applySnippetArguments(args)
-                args << featureFile.absolutePath
+                List<String> args = new CommandArgumentsBuilder(options).buildArguments(featureFile, resultsFile, junitResultsFile);
 
                 new JavaProcessLauncher(cucumberMainClass(classpath), classpath)
                         .setArgs(args)
@@ -168,84 +156,6 @@ class CucumberRunner {
         )
 
         return result
-    }
-
-    protected void applySnippetArguments(List<String> args) {
-        args << '--snippets'
-        args << options.snippets
-    }
-
-    protected void applyTagsArguments(List<String> args) {
-        if (!options.tags.isEmpty()) {
-            applyTagsToCheck(args)
-            applyTagsToIgnore(args)
-        }
-    }
-
-    private void applyTagsToCheck(List<String> args) {
-        def tagsToCheck = ''
-        def hasTags = false
-        options.tags.each {
-            if (!it.contains(TILDE)) {
-                tagsToCheck += it + ' or '
-                hasTags = true
-            }
-        }
-        if (hasTags) {
-            args << TAGS
-            args << tagsToCheck[0..-5]
-        }
-    }
-
-    private void applyTagsToIgnore(List<String> args) {
-        options.tags.each {
-            if (it.contains(TILDE)) {
-                args << TAGS
-                args << it.replaceFirst(TILDE, 'not ')
-            }
-        }
-    }
-
-    protected void applyStrictArguments(List<String> args) {
-        if (options.isStrict) {
-            args << '--strict'
-        }
-    }
-
-    protected void applyMonochromeArguments(List<String> args) {
-        if (options.isMonochrome) {
-            args << '--monochrome'
-        }
-    }
-
-    protected void applyDryRunArguments(List<String> args) {
-        if (options.isDryRun) {
-            args << '--dry-run'
-        }
-    }
-
-    protected void applyPluginArguments(List<String> args, File resultsFile, File junitResultsFile) {
-        args << PLUGIN
-        args << 'pretty'
-        args << PLUGIN
-        args << "json:${resultsFile.absolutePath}"
-        if (options.junitReport) {
-            args << PLUGIN
-            args << "junit:${junitResultsFile.absolutePath}"
-        }
-        if (!options.plugins.empty) {
-            options.plugins.each {
-                args << PLUGIN
-                args << it
-            }
-        }
-    }
-
-    protected List<String> applyGlueArguments(List<String> args) {
-        options.stepDefinitionRoots.each {
-            args << '--glue'
-            args << it
-        }
     }
 
     protected FileTree findFeatures(SourceSet sourceSet) {
